change the migrations to also include a SELECT

+++ db/migrations/postgres/000105_remove_tokens.up.sql
@@ -1,13 +1,21 @@
 DO $$
+DECLARE 
+    tokens_exist boolean := false;
 BEGIN
-WITH oauthDelete AS (
-	DELETE FROM oauthaccessdata o
+	SELECT count(o.*) != 0 INTO tokens_exist FROM oauthaccessdata o
 	WHERE NOT EXISTS (
 		SELECT p.* FROM preferences p
-		WHERE o.clientid = p.name AND o.userid = p.userid AND p.category = 'oauth_app' 
-		and p.name IS NULL
+		WHERE o.clientid = p.name AND o.userid = p.userid AND p.category = 'oauth_app'
+	);
+IF tokens_exist THEN
+	WITH oauthDelete AS (
+		DELETE FROM oauthaccessdata o
+		WHERE NOT EXISTS (
+			SELECT p.* FROM preferences p
+			WHERE o.clientid = p.name AND o.userid = p.userid AND p.category = 'oauth_app'
+		)
+		RETURNING o.token
 	)
-	RETURNING o.token
-)
-DELETE FROM sessions s WHERE s.token in (select oauthDelete.token from oauthDelete);
+	DELETE FROM sessions s WHERE s.token in (select oauthDelete.token from oauthDelete);
+END IF;
 END $$;
